Implement parametric slot feature
Introducing the concept of parametric slot:
- A new "param" argument on `trigger` and `.on` allows us to send data
through a slot for a specific parameter with `slot(<param>, <data>)`,
and reciprocally to listen to a specific parameter with
`slot.on(<param>, <data>)`.
Lazy callbacks are now called for each parameter separately: connect is
called with param as argument the first time someone connects somewhere
with `.on(<param>, ...)` and same thing for disconnect.
- is this first parameter is skipped, then a default parameter is used.
This help us when the slot has no need for such a parameter and for
retro-compatibility concerns.