Fix combineEvents and throw for duplicated events.
The generic `C[]` was forcing every eventDeclaration to have this same C
type, which is not how we intend to use combineEvents.
Therefore I am reverting the change, and updating the test so that the
case is detected.

I am also adding the proposed safeguard and throwing when encountering
duplicate events declaration in combineEvents.